+++ app/Http/Controllers/BooksController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Exports\BooksExport;
 use App\Imports\BooksImport;
+use App\Models\Order_Book;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
 use Validator;
@@ -120,8 +121,16 @@ class BooksController extends Controller
     }
 
     public function dropTrashed($id){
-        $trashedBook = Book::onlyTrashed()->find($id);
-        $trashedBook->forceDelete();
-        return redirect()->back();
+        //test if book exist in order
+        $bookOrders = Order_Book::where('book_id',$id)->get();
+
+        if($bookOrders->isEmpty()){
+            $trashedBook = Book::onlyTrashed()->find($id);
+            $trashedBook->forceDelete();
+            return redirect()->back();
+        }else{
+            $deleteProblem = 'لا يمكنك حذف الكتاب لوجود فواتير مرتبطة به';
+            return redirect()->back()->with(compact('deleteProblem'));
+        }
     }
 }
